# apps/api/openapi/compute.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AutoWaterSimu Compute API",
    "version": "0.1.0"
  },
  "servers": [{ "url": "http://localhost:8088" }],
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/compute/jobs": {
      "post": {
        "operationId": "createComputeJob",
        "parameters": [{ "name": "Idempotency-Key", "in": "header", "required": false, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ComputeJob" } } }
        },
        "responses": {
          "202": { "description": "Queued", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobSnapshot" } } } },
          "200": { "description": "Duplicate idempotency key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobSnapshot" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "get": {
        "operationId": "listComputeJobs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "job_type", "in": "query", "schema": { "type": "string" } },
          { "name": "created_after", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "created_before", "in": "query", "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": { "description": "Jobs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListJobsResponse" } } } }
        }
      }
    },
    "/api/v1/simulation-checks": {
      "post": {
        "operationId": "createSimulationCheck",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SimulationRequest" } } }
        },
        "responses": {
          "202": { "description": "Queued simulation check job", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobSnapshot" } } } },
          "200": { "description": "Duplicate simulation request idempotency key", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobSnapshot" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/process-graphs": {
      "post": {
        "operationId": "registerProcessGraph",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessGraph" } } }
        },
        "responses": {
          "201": { "description": "Registered process graph", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessGraphRecord" } } } },
          "200": { "description": "Duplicate process graph with matching payload", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProcessGraphRecord" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/simulation-inputs": {
      "post": {
        "operationId": "registerSimulationInput",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SimulationInput" } } }
        },
        "responses": {
          "201": { "description": "Registered simulation input", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SimulationInputRecord" } } } },
          "200": { "description": "Duplicate simulation input with matching payload", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SimulationInputRecord" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/scenarios": {
      "get": {
        "operationId": "listScenarios",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Scenarios", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListScenariosResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "createScenario",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenarioUpsertRequest" } } }
        },
        "responses": {
          "201": { "description": "Created scenario", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScenarioRecord" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/canvas-graphs": {
      "get": {
        "operationId": "listCanvasGraphs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "scenario_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Canvas graphs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListCanvasGraphsResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "saveCanvasGraph",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CanvasGraphSaveRequest" } } }
        },
        "responses": {
          "201": { "description": "Saved canvas graph version", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CanvasGraphRecord" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/context-snapshots": {
      "get": {
        "operationId": "listContextSnapshots",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "scenario_id", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Context snapshots", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListContextSnapshotsResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "createContextSnapshot",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContextSnapshotCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Created context snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContextSnapshotRecord" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "409": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/udm-models/templates": {
      "get": {
        "operationId": "listUdmModelTemplates",
        "parameters": [
          { "name": "tags", "in": "query", "schema": { "type": "string" } },
          { "name": "exclude_tags", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "UDM model templates", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/UDMSeedTemplateSummary" } } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/udm-models/validate": {
      "post": {
        "operationId": "validateUdmModelDefinition",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelDefinitionDraft" } } }
        },
        "responses": {
          "200": { "description": "UDM validation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMValidationResponse" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/udm-models": {
      "get": {
        "operationId": "listUdmModels",
        "parameters": [
          { "name": "skip", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "UDM models", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListUDMModelsResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "createUdmModel",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Created UDM model", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMModelDetailPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/udm-hybrid-configs": {
      "get": {
        "operationId": "listUdmHybridConfigs",
        "parameters": [
          { "name": "skip", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } }
        ],
        "responses": {
          "200": { "description": "UDM hybrid configs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListUDMHybridConfigsResponse" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "createUdmHybridConfig",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigCreateRequest" } } }
        },
        "responses": {
          "201": { "description": "Created UDM hybrid config", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UDMHybridConfigPublic" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" },
          "422": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/admin/artifacts/retention-sweep": {
      "post": {
        "operationId": "sweepArtifactRetention",
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArtifactRetentionSweepRequest" } } }
        },
        "responses": {
          "200": { "description": "Artifact retention sweep report", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ArtifactRetentionSweepReport" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/contracts/validate": {
      "post": {
        "operationId": "validateContract",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } }
        },
        "responses": {
          "200": { "description": "Validation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContractValidationResponse" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
    "/api/v1/contracts/confirm-draft": {
      "post": {
        "operationId": "confirmDraft",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } }
        },
        "responses": {
          "200": { "description": "Draft confirmation validation and persisted audit record", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContractValidationResponse" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/model-catalog": {
      "get": {
        "operationId": "listModelCatalog",
        "responses": {
          "200": { "description": "Model catalog", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelCatalog" } } } },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      },
      "post": {
        "operationId": "registerModelCatalog",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelCatalog" } } } },
        "responses": {
          "201": { "description": "Registered model catalog snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelCatalogRecord" } } } },
          "200": { "description": "Duplicate model catalog snapshot", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ModelCatalogRecord" } } } },
          "400": { "$ref": "#/components/responses/ContractError" },
          "403": { "$ref": "#/components/responses/ContractError" }
        }
      }
    },
...
    "/api/v1/model-runs": {
      "get": {
        "operationId": "listModelRuns",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "maximum": 200 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } },
          { "name": "job_id", "in": "query", "schema": { "type": "string" } },
          { "name": "model_key", "in": "query", "schema": { "type": "string" } },
          { "name": "model_version", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Model runs", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ListModelRunsResponse" } } } }
        }
      }
    },
...
    "/api/v1/workers/register": {
      "post": {
        "operationId": "registerWorker",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkerRegisterRequest" } } } },
        "responses": { "200": { "description": "Worker", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/WorkerRecord" } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": { "bearerAuth": { "type": "http", "scheme": "bearer" } },
...
    "responses": {
      "ContractError": { "description": "Contract error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContractError" } } } }
    },
    "schemas": {
      "ComputeJob": { "type": "object", "additionalProperties": true },
      "SimulationInput": { "type": "object", "additionalProperties": true },
      "SimulationRequest": {
        "type": "object",
        "required": ["schema_version", "request_id", "source_system", "requested_by", "job_type", "input_ref"],
        "properties": {
          "schema_version": { "type": "string", "enum": ["simulation_request.v1"] },
          "request_id": { "type": "string" },
          "source_system": { "type": "string" },
          "requested_by": { "type": "string" },
          "job_type": {
            "type": "string",
            "enum": [
              "simulation.material_balance.v1",
              "simulation.asm1slim.v1",
              "simulation.asm1.v1",
              "simulation.asm3.v1",
              "simulation.udm.v1"
            ]
          },
          "input_ref": {
            "type": "object",
            "additionalProperties": true,
            "properties": {
              "process_graph_id": { "type": "string" },
              "process_graph_version": { "type": "integer" },
              "simulation_input_id": { "type": "string" },
              "model_run_id": { "type": "string" },
              "simulation_input": { "type": "object", "additionalProperties": true }
            }
          },
          "external_refs": { "type": "object", "additionalProperties": true },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": true
      },
      "ProcessGraph": { "type": "object", "additionalProperties": true },
      "ProcessGraphRecord": {
        "type": "object",
        "required": ["process_graph_id", "schema_version", "version", "source_canvas_graph_id", "payload_hash", "payload", "source_system", "requested_by", "created_at"],
        "properties": {
          "process_graph_id": { "type": "string" },
          "schema_version": { "type": "string", "enum": ["process_graph.v1"] },
          "version": { "type": "integer" },
          "source_canvas_graph_id": { "type": "string" },
          "payload_hash": { "type": "string" },
          "payload": { "$ref": "#/components/schemas/ProcessGraph" },
          "source_system": { "type": "string" },
          "requested_by": { "type": "string" },
          "tenant_id": { "type": "string" },
          "project_id": { "type": "string" },
          "site_id": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "SimulationInputRecord": {
        "type": "object",
        "required": ["simulation_input_id", "schema_version", "job_type", "process_graph_id", "process_graph_version", "payload_hash", "payload", "source_system", "requested_by", "created_at"],
        "properties": {
          "simulation_input_id": { "type": "string" },
          "schema_version": { "type": "string", "enum": ["simulation_input.v1"] },
          "job_type": { "type": "string" },
          "process_graph_id": { "type": "string" },
          "process_graph_version": { "type": "integer" },
          "payload_hash": { "type": "string" },
          "payload": { "$ref": "#/components/schemas/SimulationInput" },
          "source_system": { "type": "string" },
          "requested_by": { "type": "string" },
          "tenant_id": { "type": "string" },
          "project_id": { "type": "string" },
          "site_id": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ScenarioRecord": {
        "type": "object",
        "required": ["scenario_id", "name", "model_family", "status", "version", "source_system", "requested_by", "created_at", "updated_at"],
        "properties": {
          "scenario_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "model_family": { "type": "string" },
          "status": { "type": "string" },
          "version": { "type": "integer" },
          "current_canvas_graph_id": { "type": "string" },
          "current_canvas_graph_version": { "type": "integer" },
          "published_process_graph_id": { "type": "string" },
          "published_process_graph_version": { "type": "integer" },
          "current_simulation_input_id": { "type": "string" },
          "context_snapshot_id": { "type": "string" },
          "source_scenario_id": { "type": "string" },
          "last_job_id": { "type": "string" },
          "source_system": { "type": "string" },
          "requested_by": { "type": "string" },
          "tenant_id": { "type": "string" },
          "project_id": { "type": "string" },
          "site_id": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "archived_at": { "type": "string", "format": "date-time" }
        }
      },
      "ListScenariosResponse": {
        "type": "object",
        "required": ["items", "total_estimate"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ScenarioRecord" } },
          "next_cursor": { "type": "string" },
          "total_estimate": { "type": "integer" }
        }
      },
      "ScenarioUpsertRequest": {
        "type": "object",
        "properties": {
          "scenario_id": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "model_family": { "type": "string" },
          "status": { "type": "string" },
          "current_canvas_graph_id": { "type": "string" },
          "current_canvas_graph_version": { "type": "integer" },
          "published_process_graph_id": { "type": "string" },
          "published_process_graph_version": { "type": "integer" },
          "current_simulation_input_id": { "type": "string" },
          "context_snapshot_id": { "type": "string" },
          "last_job_id": { "type": "string" },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": false
      },
...
      "CanvasGraphRecord": {
        "type": "object",
        "required": ["graph_id", "schema_version", "name", "version", "payload_hash", "payload", "source_system", "requested_by", "created_at"],
        "properties": {
          "graph_id": { "type": "string" },
          "scenario_id": { "type": "string" },
          "schema_version": { "type": "string", "enum": ["canvas_graph.v1"] },
          "name": { "type": "string" },
          "version": { "type": "integer" },
          "payload_hash": { "type": "string" },
          "payload": { "$ref": "#/components/schemas/CanvasGraph" },
          "source_system": { "type": "string" },
          "requested_by": { "type": "string" },
          "tenant_id": { "type": "string" },
          "project_id": { "type": "string" },
          "site_id": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "created_at": { "type": "string", "format": "date-time" },
          "archived_at": { "type": "string", "format": "date-time" }
        }
      },
      "ListCanvasGraphsResponse": {
        "type": "object",
        "required": ["items", "total_estimate"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/CanvasGraphRecord" } },
          "next_cursor": { "type": "string" },
          "total_estimate": { "type": "integer" }
        }
      },
      "CanvasGraphSaveRequest": {
        "type": "object",
        "required": ["canvas_graph"],
        "properties": {
          "scenario_id": { "type": "string" },
          "graph_id": { "type": "string" },
          "name": { "type": "string" },
          "canvas_graph": { "$ref": "#/components/schemas/CanvasGraph" },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": false
      },
...
      "ContextSnapshotRecord": {
        "type": "object",
        "required": ["context_snapshot_id", "schema_version", "source_system", "captured_at", "payload_hash", "payload", "created_at"],
        "properties": {
          "context_snapshot_id": { "type": "string" },
          "scenario_id": { "type": "string" },
          "schema_version": { "type": "string", "enum": ["context_snapshot.v1"] },
          "source_system": { "type": "string" },
          "captured_at": { "type": "string", "format": "date-time" },
          "payload_hash": { "type": "string" },
          "payload": { "type": "object", "additionalProperties": true },
          "tenant_id": { "type": "string" },
          "project_id": { "type": "string" },
          "site_id": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ListContextSnapshotsResponse": {
        "type": "object",
        "required": ["items", "total_estimate"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ContextSnapshotRecord" } },
          "next_cursor": { "type": "string" },
          "total_estimate": { "type": "integer" }
        }
      },
      "ContextSnapshotCreateRequest": {
        "type": "object",
        "properties": {
          "context_snapshot_id": { "type": "string" },
          "scenario_id": { "type": "string" },
          "source_system": { "type": "string" },
          "context": { "type": "object", "additionalProperties": true },
          "metadata": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": false
      },
      "UDMModelDefinitionDraft": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "components": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "parameters": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "processes": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "meta": { "type": "object", "additionalProperties": true },
          "validation_mode": { "type": "string" }
        }
      },
      "UDMModelCreateRequest": {
        "allOf": [
          { "$ref": "#/components/schemas/UDMModelDefinitionDraft" },
          {
            "type": "object",
            "properties": {
              "seed_source": { "type": "string" },
              "source_system": { "type": "string" },
              "metadata": { "type": "object", "additionalProperties": true }
            }
          }
        ]
      },
...
      "UDMSeedTemplateSummary": {
        "type": "object",
        "required": ["key", "name"],
        "properties": {
          "key": { "type": "string" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } },
          "components_count": { "type": "integer" },
          "processes_count": { "type": "integer" },
          "parameters_count": { "type": "integer" }
        }
      },
...
      "UDMValidationResponse": {
        "type": "object",
        "required": ["ok", "errors", "warnings", "extracted_parameters", "continuity_checks"],
        "properties": {
          "ok": { "type": "boolean" },
          "errors": { "type": "array", "items": { "$ref": "#/components/schemas/UDMValidationIssue" } },
          "warnings": { "type": "array", "items": { "$ref": "#/components/schemas/UDMValidationIssue" } },
          "extracted_parameters": { "type": "array", "items": { "type": "string" } },
          "continuity_checks": { "type": "array", "items": { "$ref": "#/components/schemas/UDMContinuityCheckItem" } }
        }
      },
...
      "UDMModelDetailPublic": {
        "allOf": [
          { "$ref": "#/components/schemas/UDMModelPublic" },
          {
            "type": "object",
            "required": ["versions"],
            "properties": {
              "latest_version": { "$ref": "#/components/schemas/UDMModelVersionPublic" },
              "versions": { "type": "array", "items": { "$ref": "#/components/schemas/UDMModelVersionPublic" } }
            }
          }
        ]
      },
      "ListUDMModelsResponse": {
        "type": "object",
        "required": ["data", "count"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/UDMModelPublic" } },
          "count": { "type": "integer" }
        }
      },
      "UDMHybridConfigCreateRequest": {
        "type": "object",
        "required": ["name", "hybrid_config"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "hybrid_config": { "type": "object", "additionalProperties": true },
          "source_system": { "type": "string" },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
...
      "UDMHybridConfigPublic": {
        "type": "object",
        "required": ["id", "owner_id", "created_at", "updated_at", "name", "hybrid_config"],
        "properties": {
          "id": { "type": "string" },
          "owner_id": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "hybrid_config": { "type": "object", "additionalProperties": true },
          "parameter_hash": { "type": "string" }
        }
      },
      "ListUDMHybridConfigsResponse": {
        "type": "object",
        "required": ["data", "count"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/UDMHybridConfigPublic" } },
          "count": { "type": "integer" }
        }
      },
...
      "ArtifactRetentionSweepRequest": {
        "type": "object",
        "properties": {
          "dry_run": { "type": "boolean", "default": true },
          "limit": { "type": "integer", "minimum": 0, "maximum": 1000 }
        }
      },
...
      "JobSnapshot": {
        "type": "object",
        "required": ["job", "artifacts", "event_count"],
        "properties": {
          "job": { "$ref": "#/components/schemas/JobRecord" },
          "artifacts": { "type": "array", "items": { "$ref": "#/components/schemas/ArtifactRecord" } },
          "event_count": { "type": "integer" }
        }
      },
      "ListJobsResponse": {
        "type": "object",
        "required": ["items", "total_estimate"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/JobSnapshot" } },
          "next_cursor": { "type": "string" },
          "total_estimate": { "type": "integer" }
        }
      },
      "ListModelRunsResponse": {
        "type": "object",
        "required": ["items", "total_estimate"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/ModelRun" } },
          "next_cursor": { "type": "string" },
          "total_estimate": { "type": "integer" }
        }
      },
...
      "WorkerRecord": { "type": "object", "additionalProperties": true },
...
    }
  }
}